refactor: move waitForEditorsToMount logic to EditorMosaic
